backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
aiofiles
requests
replicate
gTTS
TTS
opencv-python
librosa
gdown
scipy
imageio[ffmpeg]
